pyproject.toml
[project]
name = "component-databricks-job-runner"
dynamic = ["version"]
requires-python = ">=3.13"
dependencies = [
    "flake8>=7.3.0",
    "keboola-component>=1.6.13",
    "keboola-http-client>=1.0.1",
    "mock>=4.0.3",
    "pydantic>=2.11.9",
]
